+++ utilities/blob_db/blob_db_test.cc
@@ -18,6 +18,7 @@
 #include "util/cast_util.h"
 #include "util/fault_injection_test_env.h"
 #include "util/random.h"
+#include "util/sst_file_manager_impl.h"
 #include "util/string_util.h"
 #include "util/sync_point.h"
 #include "util/testharness.h"
@@ -688,11 +689,31 @@ TEST_F(BlobDBTest, DISABLED_GCOldestSimpleBlobFileWhenOutOfSpace) {
 
 TEST_F(BlobDBTest, ReadWhileGC) {
   // run the same test for Get(), MultiGet() and Iterator each.
+  std::shared_ptr<SstFileManager> sst_file_manager(
+      NewSstFileManager(mock_env_.get()));
+  sst_file_manager->SetDeleteRateBytesPerSecond(1);
+  SstFileManagerImpl *sfm =
+      static_cast<SstFileManagerImpl *>(sst_file_manager.get());
   for (int i = 0; i < 2; i++) {
     BlobDBOptions bdb_options;
     bdb_options.min_blob_size = 0;
     bdb_options.disable_background_tasks = true;
-    Open(bdb_options);
+
+    Options db_options;
+
+    int files_deleted_directly = 0;
+    int files_scheduled_to_delete = 0;
+    rocksdb::SyncPoint::GetInstance()->SetCallBack(
+        "SstFileManagerImpl::ScheduleFileDeletion",
+        [&](void * /*arg*/) { files_scheduled_to_delete++; });
+    rocksdb::SyncPoint::GetInstance()->SetCallBack(
+        "DeleteScheduler::DeleteFile",
+        [&](void * /*arg*/) { files_deleted_directly++; });
+    if (i == 0) {
+      db_options.sst_file_manager = sst_file_manager;
+    }
+
+    Open(bdb_options, db_options);
     blob_db_->Put(WriteOptions(), "foo", "bar");
     auto blob_files = blob_db_impl()->TEST_GetBlobFiles();
     ASSERT_EQ(1, blob_files.size());
@@ -749,16 +770,29 @@ TEST_F(BlobDBTest, ReadWhileGC) {
     ASSERT_EQ(bfile_number, obsolete_files[0]->BlobFileNumber());
     TEST_SYNC_POINT("BlobDBTest::ReadWhileGC:2");
     reader.join();
-    SyncPoint::GetInstance()->DisableProcessing();
 
     // The file is deleted this time
     blob_db_impl()->TEST_DeleteObsoleteFiles();
+    SyncPoint::GetInstance()->DisableProcessing();
     blob_files = blob_db_impl()->TEST_GetBlobFiles();
     ASSERT_EQ(1, blob_files.size());
     ASSERT_NE(bfile_number, blob_files[0]->BlobFileNumber());
     ASSERT_EQ(0, blob_db_impl()->TEST_GetObsoleteFiles().size());
     VerifyDB({{"foo", "bar"}});
+    // Even if SSTFileManager is not set, DB is creating a dummy one.
+    ASSERT_EQ(1, files_scheduled_to_delete);
+    if (i == 0) {
+      // Delete is rate limited
+      ASSERT_EQ(0, files_deleted_directly);
+    } else {
+      ASSERT_EQ(1, files_deleted_directly);
+    }
     Destroy();
+    if (i == 0) {
+      // Make sure that DestroyBlobDB() also goes through delete scheduler.
+      ASSERT_GE(2, files_scheduled_to_delete);
+    }
+    sfm->WaitForEmptyTrash();
   }
 }